security for video. add videos in user

+++ controllers/videoController.js
@@ -39,9 +39,10 @@ export const postUpload = async (req, res) => {
         fileUrl: path,
         title,
         description,
+        creator: req.user.id,
     });
-    console.log(typeof newVideo);
-    //TO DO : upload and save video
+    req.user.videos.push(newVideo.id);
+    req.user.save();
     res.redirect(routes.videoDetail(newVideo.id));
 };
 
@@ -51,7 +52,8 @@ export const videoDetail = async (req, res) => {
     } = req;
 
     try {
-        const video = await Video.findById(id);
+        const video = await Video.findById(id).populate("creator");
+        console.log(video);
         res.render("videoDetail", { pageTitle: video.title, video });
     } catch (error) {
         res.redirect(routes.home);
@@ -64,6 +66,9 @@ export const getEditVideo = async (req, res) => {
 
     try {
         const video = await Video.findById(id);
+        if (video.creator !== req.user.id) {
+            throw Error();
+        }
         res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
     } catch (error) {
         res.redirect(routes.home);
@@ -99,6 +104,11 @@ export const deleteVideo = async (req, res) => {
     console.log(id);
 
     try {
+        const video = await Video.findById(id);
+        if (video.creator !== req.user.id) {
+            throw Error();
+        }
+
         await Video.findOneAndRemove({ _id: id });
     } catch (error) {
         console.log(error);
